requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.22.0

# Data Reading
openpyxl>=3.1.0
xlrd>=2.0.1

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0

# Utils
python-dateutil>=2.8.0

# MLOps & Model Serving
imbalanced-learn>=0.12.0
joblib>=1.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Notebook Execution
nbformat>=5.9.0

# GCP
google-cloud-storage>=2.10.0
google-auth>=2.22.0